requirements.txt
# Python 3.11 targeted requirements
pywebview>=4.4   ## For desktop application GUI
streamlit>=1.17.0  ## For web application GUI
requests>=2.32.5  ## For HTTP requests
datetime>=5.5  ## For date and time manipulation
pyside6>=6.10.0  ## For GUI components
qtpy>=2.4.1  ## For Qt abstraction layer
opencv-python>=3.4.0.14  ## For image processing
Pillow>=10.0.0  ## For image handling
pytesseract>=0.3.10  ## For OCR functionality
httpx>=0.28.1  ## For async HTTP requests
Unidecode>=1.4.0  ## For text normalization
EasyOCR>=1.7.2  ## For optical character recognition
pybase64>=1.4.2  ## For base64 encoding/decoding
numpy>=2.2.6  ## For numerical operations
pandas>=2.2.3  ## For data manipulation and analysis